+++ tests/testthat/test-cbind.R
@@ -17,4 +17,13 @@ test_that( "cbind uses shallow copies", {
   expect_equal( dfloc(df2), dfloc(df)[names(df2)] )
 })
 
-
+test_that( "bind_cols produces a tbl_df (#779)", {
+  df1 <- data.frame(a = letters)
+  df2 <- data.frame(A = LETTERS)
+    
+  expect_is( bind_cols(df1, df2), "tbl_df")
+  expect_is( bind_cols(df1, tbl_df(df2)), "tbl_df")
+  expect_is( bind_cols(tbl_df(df1), df2), "tbl_df")
+  expect_is( bind_cols(tbl_df(df1), tbl_df(df2)), "tbl_df")
+  
+})